+++ lib/src/06_auth/FauiAuthState.dart
@@ -1,5 +1,0 @@
-import 'package:faui/FauiUser.dart';
-
-class FauiAuthState {
-  static FauiUser user;
-}