+++ lib/src/09_utility/HttpMethod.dart
@@ -1,5 +1,0 @@
-enum HttpMethod {
-  patch,
-  get,
-  post,
-}