+++ lib/src/10_auth/auth_state.dart
@@ -1,0 +1,5 @@
+import 'package:faui/src/90_model/faui_user.dart';
+
+class FauiAuthState {
+  static FauiUser user;
+}